chore(hub-protocol): split canonical.rs into a module tree (#140)
canonical.rs grew to ~1200 lines hosting four distinct concerns
sharing a namespace (const-fn postcard primitives, schema +
kind serializers, KindLabels sidecar, ADR-0033 InputsRecord
encoders). Split in place into canonical/{mod,primitives,schema,
labels,inputs}.rs with cross-sibling helpers marked pub(super).
The module re-exports every pub item so canonical::* callers
(aether-mail-derive, aether-substrate, aether-hub, tests) compile
unchanged — no downstream edits required.

Closes #140.
